Move symbolic domain utilities out of global tmp pass into seperate module
